fix: fixing infinite fetches
