+++ app/academia/page.tsx
@@ -5,38 +5,16 @@ import Marks from "./components/Marks";
 import Attendance from "./components/Attendance";
 import Loading from "@/components/States/Loading";
 import { useEffect, useState } from "react";
-// ---
 import { useUser } from "@/provider/UserProvider";
-import { useAttendance } from "@/provider/AttendanceProvider";
-import { useDay } from "@/provider/DayProvider";
-import { useMarks } from "@/provider/MarksProvider";
-import { useTimetable } from "@/provider/TimetableProvider";
 
-function useMutateAll() {
-  const { mutate: mutateAttendance } = useAttendance();
-  const { mutate: mutateDay } = useDay();
-  const { mutate: mutateMarks } = useMarks();
-  const { mutate: mutateTimetable } = useTimetable();
-  const { mutate: mutateUser } = useUser();
-
-  return () => {
-    mutateUser();
-    mutateAttendance();
-    mutateDay();
-    mutateMarks();
-    mutateTimetable();
-  };
-}
 
 export default function Academia() {
   const { user, isLoading } = useUser();
 
   const [mounted, setMounted] = useState(false);
-  const mutateAll = useMutateAll();
 
   useEffect(() => {
     setMounted(true);
-    mutateAll();
   }, []);
 
   if (!mounted) return null;
